softskillspace-develop/requirements.txt
boto3
django-ckeditor==6.4
channels-redis==3.4.0
daphne==3.0.2
dj_database_url==0.5.0
django-allauth==0.49.0
django-auto-prefetch==1.1.0
django-ckeditor==6.4.0
django-debug-toolbar==3.2.4
django-redis==5.2.0
django-resized==0.3.11
django-storages==1.12.3
django==3.2.13
dropbox==10.4.1
pillow==9.0.1
psycopg2-binary==2.9.3
redis==4.2.0
stripe==2.74.0
twilio==7.7.1
whitenoise==6.0.0
django-db-connection-pool==1.1.0
